move test that was in wrong file
